better implementation to transform implicit 'else' branch of multi cycle 'if' statement, and corrects potential bug.
We now create an iterable of branches, adding a dummy branch if the 'if'
has no 'else'. This has the effect of registering any peek done in the
different conditions as reads. Until now, the 'else' branch could be
taken if there was no data available, even in the absence of any
'available' test in the branches before it.